# docs/swagger.yaml
basePath: /
definitions:
...
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.Ciudadano:
    properties:
      alias:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      tenant_id:
        type: integer
    type: object
  entities.CiudadanoListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.Ciudadano'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.CiudadanoMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.CiudadanoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.Ciudadano'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.CreateCiudadanoRequest:
    properties:
      alias:
        type: string
      email:
        type: string
      fcm_token:
        type: string
      password:
        type: string
    required:
    - alias
    - email
    - fcm_token
    - password
    type: object
...
  entities.CreateHistorialAsignacionCamionRequest:
    properties:
      fecha_asignacion:
        type: string
      fecha_baja:
        type: string
      id_camion:
        type: integer
      id_chofer:
        type: integer
      id_historial:
        type: integer
    type: object
...
  entities.HistorialAsignacionCamion:
    properties:
      created_at:
        type: string
      eliminado:
        type: boolean
      fecha_asignacion:
        type: string
      fecha_baja:
        type: string
      id_camion:
        type: integer
      id_chofer:
        type: integer
      id_historial:
        type: integer
      updated_at:
        type: string
    type: object
  entities.HistorialAsignacionCamionListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.HistorialAsignacionCamion'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.HistorialAsignacionCamionMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.HistorialAsignacionCamionResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.HistorialAsignacionCamion'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.UpdateCiudadanoRequest:
    properties:
      alias:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  entities.UpdateHistorialAsignacionCamionRequest:
    properties:
      fecha_asignacion:
        type: string
      fecha_baja:
        type: string
      id_camion:
        type: integer
      id_chofer:
        type: integer
      id_historial:
        type: integer
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/ciudadanos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar ciudadanos
      tags:
      - Ciudadano
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateCiudadanoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.CiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Crear ciudadano
      tags:
      - Ciudadano
  /api/ciudadanos/{id}:
    delete:
      parameters:
      - description: ID del ciudadano
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar ciudadano
      tags:
      - Ciudadano
    get:
      parameters:
      - description: ID del ciudadano
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener ciudadano por ID
      tags:
      - Ciudadano
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID del ciudadano
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateCiudadanoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.CiudadanoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar ciudadano
      tags:
      - Ciudadano
...
  /api/historial-asignacion/{id}:
    delete:
      parameters:
      - description: ID historial
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.HistorialAsignacionCamionMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar historial de asignación
      tags:
      - HistorialAsignacion
    get:
      parameters:
      - description: ID historial
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.HistorialAsignacionCamionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener historial por ID
      tags:
      - HistorialAsignacion
    put:
      consumes:
      - application/json
      parameters:
      - description: ID historial
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateHistorialAsignacionCamionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.HistorialAsignacionCamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar historial de asignación
      tags:
      - HistorialAsignacion
...
  /api/ruta-camion/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.HistorialAsignacionCamionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar rutas-camión
      tags:
      - RutaCamion
    post:
      consumes:
      - application/json
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateHistorialAsignacionCamionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.HistorialAsignacionCamionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear ruta-camión
      tags:
      - RutaCamion
...
  /api/ruta-camion/exists/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.HistorialAsignacionCamionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verificar existencia ruta-camión
      tags:
      - RutaCamion
...
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"